API docs without RequestState
