+++ package.json
@@ -13,7 +13,7 @@
     "docs": "typedoc src/index.ts",
     "lint": "tsc --noEmit && eslint --report-unused-disable-directives && prettier --check src",
     "prepare": "pnpm run build && husky",
-    "tsoa": "tsoa spec-and-routes && tsc && redocly build-docs tsoa_build/swagger.json"
+    "tsoa": "./scripts/tsoa.sh"
   },
   "repository": {
     "type": "git",
