+++ src/request_state.ts
@@ -18,16 +18,3 @@ export type RequestState<T> =
   | LoadingRequestState
   | FailedRequestState
   | SuccessRequestState<T>;
-
-export function mapToRequestState<ResultT>(
-  result?: ResultT,
-  error?: Error,
-): RequestState<ResultT> {
-  if (error) {
-    return { state: "failed", err: error };
-  }
-  if (result !== undefined) {
-    return { state: "success", data: result };
-  }
-  return { state: "empty" };
-}
